extract interface into its own file
